requirements.txt
gql[all]==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
typing_extensions==4.15.0
